requirement.txt
fastapi
pydantic
pydantic[Email]
sqlalchemy
pyjwt
psycopg2-binary
pytest
httpx
python-multipart